dom/bindings/test/test_callback_default_thisval.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=957929
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 957929</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 957929 **/
  SimpleTest.waitForExplicitFinish();

  function f() {
    "use strict";
    is(this, undefined, "Should have undefined this value");
    SimpleTest.finish();
  }

  addLoadEvent(function() {
    requestAnimationFrame(f);
  });
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=957929">Mozilla Bug 957929</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
